Struct seek_bufread::BufReader

pub struct BufReader<R> { /* fields omitted */ }

The BufReader struct adds buffering to any reader.

It can be excessively inefficient to work directly with a Read instance. For example, every call to read on TcpStream results in a system call. A BufReader performs large, infrequent reads on the underlying Read and maintains an in-memory buffer of the results.

Examples

use std::io::prelude::*;
use std::fs::File;
use seek_bufread::BufReader;

let mut f = try!(File::open("log.txt"));
let mut reader = BufReader::new(f);

let mut line = String::new();
let len = try!(reader.read_line(&mut line));
println!("First line is {} bytes long", len);Run

Methods

impl<R: Read + Seek> BufReader<R>

fn new(inner: R) -> BufReader<R>

Creates a new BufReader with a default buffer capacity (8192 bytes).

Examples

use std::fs::File;
use seek_bufread::BufReader;

let mut f = try!(File::open("log.txt"));
let mut reader = BufReader::new(f);Run

fn with_capacity(cap: usize, inner: R) -> BufReader<R>

Creates a new BufReader with the specified buffer capacity.

Examples

Creating a buffer with ten bytes of capacity:

use std::fs::File;
use seek_bufread::BufReader;

let mut f = try!(File::open("log.txt"));
let mut reader = BufReader::with_capacity(10, f);Run

fn position(&self) -> u64

Returns the absolute file pointer position.

fn capacity(&self) -> usize

Returns the total buffer capacity.

fn available(&self) -> usize

Returns the current number of remaining bytes available in the buffer.

fn into_inner(self) -> Result<R>

Consumes self, synchronizes the inner reader position and returns the inner reader.